Corrected minor merging issues

+++ src/com/upb/taxbilling/view/ExportBill.java
@@ -2,23 +2,16 @@ package com.upb.taxbilling.view;
 
 import java.io.File;
 import java.io.FileOutputStream;
-import java.io.FileWriter;
 import java.io.OutputStreamWriter;
-import java.io.PrintWriter;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.Locale;
 
-import com.upb.taxbilling.R;
-import com.upb.taxbilling.model.data.Bill;
-import com.upb.taxbilling.view.billtable.BillTableFragment;
-
 import android.app.Fragment;
 import android.os.Bundle;
 import android.os.Environment;
-import android.provider.Telephony.Sms.Conversations;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
@@ -26,6 +19,9 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.Toast;
+
+import com.upb.taxbilling.R;
+import com.upb.taxbilling.model.data.Bill;
 
 /**
  * The fragment where the information about a User and Bill is export to a file
@@ -186,7 +182,7 @@ public class ExportBill extends Fragment{
 		{
 			ArrayBill.add(Integer.toString(ArrayBillData.get(i).getNit())+"|"
 		              +Integer.toString(ArrayBillData.get(i).getBillNumber())+"|"
-		              +Integer.toString(ArrayBillData.get(i).getAutorizationNumber())+"|"
+		              +Long.toString(ArrayBillData.get(i).getAutorizationNumber())+"|"
 		              +df.format(ArrayBillData.get(i).getEmissionDate())+"|"
 				      +Double.toString(ArrayBillData.get(i).getAmount())+"|"
 				      +ArrayBillData.get(i).getControlCode());
